feat(iotmonitoring.concretecuring): add concrete curing sensor context facade implementation which provides a simplified integration surface for other bounded contexts that need concrete curing operations.

+++ backend/src/main/java/com/kipu/backend/iotmonitoring/concretecuring/interfaces/acl/ConcreteCuringSensorContextFacade.java
@@ -25,7 +25,7 @@ public interface ConcreteCuringSensorContextFacade {
             String location,
             Double temperatureReading,
             String temperatureUnit,
-            Double humidityPercentage,
+            Integer humidityPercentage,
             Double temperatureLimit
     );
